// SRC/Manifest/Saras.eMarkingClient/src/app/model/project/standardisation/std-one/recommendation/recommendation-model.ts
export interface IRecommedData {
    ProjectId: number;
    ScriptId: number;
    ScriptName: string;
    QigId: number;
    QigName: string;
    IsViewMode: boolean | undefined;
    Markedby?: any;
    Workflowid?: number;

}

export interface IBandingScriptResponse {
    element: {};
    ScriptId: number;
    ProjectQnsId: number;
    TotalNoOfQuestions: number;
    QuestionCode: string;
    QuestionText: string;
    QuestionOrder: number;
    ResponseText: string;
    RecommendedBand?: IBanding;
    Bands: IBanding[];
    IsActive: boolean;
    IsQigLevel: boolean;
    IsMarkSchemeTagged: boolean;
    IsNullResponse:boolean;
}

export interface IBanding {
    BandId: number;
    BandName: string;
    BandCode: string;
    IsSelected: boolean;
    BandFrom: number;
    BandTo: number;
}

export class UnrecommandedScript {
  ScriptId!: number;
  Qigid!: number;
}